Add contactFields property to Group
This commit adds the contactFields array to the Group entity. Currently, it's
implemented to support e-mail contact only in the first version. However, by
implementing this as an associative array of label => uri, this can be easily
extended later on. The form and group overview page have been updated
respectively. The activity page has a contact button added.